Gate the ZODL Slipstream engine behind a cargo feature
slipstream-core (zodl-slipstream, AGPL-3.0-only) and tokio become
optional dependencies behind a new `slipstream` feature; `gpu` now
implies it. The whole zcashlc_slipstream_* FFI surface moves verbatim
from lib.rs into rust/src/slipstream_ffi.rs, gated together with
retained_marks (its flush compensator) by one cfg per mod line, so the
file boundary doubles as the licensing audit boundary. The ffi.rs
recovery-hold helper cluster that only the engine surface calls is
gated the same way.

cbindgen now emits the slipstream declarations wrapped in
serves both artifact variants; packaging specializes it with unifdef.

Verified: cargo check/test with and without --features slipstream
(126 and 162 tests passing), zodl absent from the default-feature
dependency tree, and unifdef -U strips the header to zero slipstream
references.
